Refactor HTML export for family tree generation
The process of generating HTML for family trees has been refactored. Instead of generating the structure in the export file, it is now built in the htmlexport.c. To align with this, relevant files and file parts have been removed or updated. This change enhances efficiency by consolidating the creation process into the source file.